Fix clang-tidy and buildifier warnings.
