+++ src/ir/types/proto/BUILD
@@ -14,8 +14,6 @@
 # limitations under the License.
 #
 #-------------------------------------------------------------------------------
-load("//build_defs:test_helpers.bzl", "extracted_datalog_string_test")
-
 package(
     default_visibility = ["//src:__subpackages__"],
     features = ["layering_check"],
